Fix wrongly moved models
