+++ src/main/java/no/oslomet/aaas/model/AnonymizationResultPayload.java
@@ -1,4 +1,6 @@
-package no.oslomet.aaas.model.analytics;
+package no.oslomet.aaas.model;
+
+import no.oslomet.aaas.model.analytics.RiskProfile;
 
 
 /***
