+++ src/main/java/no/oslomet/aaas/service/AnonymizationService.java
@@ -3,8 +3,8 @@ package no.oslomet.aaas.service;
 import no.oslomet.aaas.analyzer.Analyzer;
 import no.oslomet.aaas.anonymizer.Anonymizer;
 import no.oslomet.aaas.model.*;
-import no.oslomet.aaas.model.analytics.AnonymizationResultPayload;
-import no.oslomet.aaas.model.analytics.AnonymizeResult;
+import no.oslomet.aaas.model.AnonymizationResultPayload;
+import no.oslomet.aaas.model.AnonymizeResult;
 import no.oslomet.aaas.model.analytics.RiskProfile;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
